Extract context classes into separate files for better organization
